+++ tests/core/test_core.py
@@ -35,7 +35,7 @@ def fixture_candles_31st(candles):
 def test_data(candles_30, expected_ema):
     test = EMA(candles=candles_30, input_value="close")
     test.calculate()
-    assert pytest.approx(test.as_list) == expected_ema[0:30]
+    assert pytest.approx(test.as_list()) == expected_ema[0:30]
 
 
 def test_data_append(candles_30, candles_31st, expected_ema):
@@ -44,4 +44,4 @@ def test_data_append(candles_30, candles_31st, expected_ema):
     candles_30.append(candles_31st)
     test.calculate()
 
-    assert pytest.approx(test.as_list) == expected_ema[0:31]
+    assert pytest.approx(test.as_list()) == expected_ema[0:31]