+++ tests/indicators/test_amoprh_patterns.py
@@ -112,10 +112,10 @@ class TestPatterns:
     def test_doji(self, candles, expected_doji):
         test = indicators.Amorph(analysis=patterns.doji, candles=candles)
         test.calculate()
-        assert self.verify(test.as_list, expected_doji)
+        assert self.verify(test.as_list(), expected_doji)
 
     @pytest.mark.usefixtures("candles", "expected_hammer")
     def test_hammer(self, candles, expected_hammer):
         test = indicators.Amorph(analysis=patterns.hammer, candles=candles)
         test.calculate()
-        assert self.verify(test.as_list, expected_hammer)
+        assert self.verify(test.as_list(), expected_hammer)